backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
bcrypt==4.1.3
passlib>=1.7.4
python-jose>=3.3.0
requests>=2.31.0
python-multipart>=0.0.9
supabase>=2.3.0
postgrest>=0.14.0

# Dependências opcionais para testes e desenvolvimento
pytest>=8.0.0
black>=24.1.1
isort>=5.13.2
